Making the Service into a Component and manager of Gatway objects.
The idea here is as follows... Since the starting point of the
application is the loading of the web service object by the Hessian
servlet, this is where we have to setup (instantiate) the
DataNodeManager.  By making the Service a "Component", the service can
then register itself with the data node manager it instantiated and
thus be able to capture events; Namely the events that inform it that
the gateway proxy objects (instances of BasicGateway created in and by
the data node manager) have been registered.  The service captures
these events and holds onto the Gateway objects and manages them.

The other change in this commit has to do with how the service manages
the Gateway proxy objects it has collected and how the Gateway proxy
objects themselves set their internal state.  The Service sets up a
timer that uses the Gateway proxy objects to call "ping" on the
gateway in quesetion.  If the ping comes back false or does not come
back at all (times out because of not being able to connect), the
Gateway proxy object sets its internal state to unavailable
(isAvailable=false). The idea is that if communication could not take
place then the object should balk at further communications by having
a guard at the top of each method that wraps an external call.  The
Service will place the "unavailable" Gateway proxy objects into a
separate data structure and will continue to attempt to call ping, in
the event that the endpoint will respond with true and thus change the
state of the gateway proxy.  If the object becomes available then it
will be transfered to the active data structure.  There service itself
also has state.  It will refuse incoming rpc calls if it has no
available gateway proxy objects.  This is a quick first cut at this
logic. As long as there is at least one available gateway proxy the
service's state will be "available" and it will accept ingress calls.
This should be changed to be more discerning, as there could be a
single good gateway proxy that keeps the door open for another outside
Mallory to DOS the system. A better way would be to check where each
call is coming from and then make sure it matches one of the gateway
proxies that we have handles to... But that would require directly
being a subclass of the HessianServlet so that we would have more
access to the input flow (perhaps). For now we just want to have at
least this rudimantary connection management.