Setup of initial round trip node to gway test
Put in faux behavior in strawman gateway to respond to registration
requests.  Coded up a periodic call to register from the data node to
the gateway.  Instantiating a single gateway connection stub object to
test.  The goal here is to get the p to p logic established.  This
includes the boostrapping from the DataNodeService and Join events
handling.