Good place... Things work! Round trip testing of boostrap and rpc.
The bug from before was surmounted by changing the ordering of the
boostraping such that the service registersters itself to the data
node manager then initializes the data node manager so it would be
ready to listen and recieve join requests.  This was good but the real
solution was in the active fetching of gateways from the data node
manager.  Though I didn't want to change this from protected to
private, it is pretty much what would have to be done for components
needing resources that have been loaded before them. (see logic in the
ESGDataNodeServiceImpl's registration timer.