Closer. There seems to be problem with shifting lists...
There seems to e a problem with moving nodes from then unavailable
list to the available list.  I am not sure what it is yet. So far all
the data coming from the Gateway object (BasicGateway) is correct and
expected.  The code in the timers in tyhe ESGDataNodeServiceImpl seem
to be not performing properly with regards to moving state lists.
(going home now, my wife will kill me) The price we pay for striving
for perfection ;-)