New ConnectionManager works... (just as well as previous scenario)
We didn't loose ground but now we have to think a bit more about the
communiation path from the gateway proxy objects over to the
connection manager and then over the the service.  The idea is that
if there is a problem with the gateway proxy (no more communication)
then we want to let the connection manager know so he can be called by
the service to know if the service should shut down responding.

It seems as though the next logical step would be to make the
connection manager a listener to the proxy objects... right?

I am mulling this over, as I am trying to keep this as simple as
possible and no simpler.  This is at odds with my strict sensibilities
of separation of concerns.  There should be a single location or at
least package for dealing with communication semantics.  There should
certainly be a proxy object to represent and fascilitate egress
communication. The real issue is that the service, must be it's own
beast because of the use of Hessian and services (sort of).  Okay I
convinced myself...

TODO: I am going to add a listener for the proxy objects. It also is
cleaner in some ways because it makes the system passive, no
polling!!!! :-)... if a call fails the proxy announces this to the
connection manager and at that time the manager does the write book
keeping and thus can be a resource for others who what to know the
overall state of connections (i.e. the service).