Added JAXB libraries and build target etc.
I want to add the ability for this project to parse the "catalog" that
we deal with.  At the moment the idea is to pass the raw string of the
catalog as the payload on the remote event and then parse it into
objects using JAXB and then repackage the data as objects into a local
event objects to dispatch through our 'state machine' ;-).

Being able to parse and re-represent XML is going to be potentially
quite useful... IMHO.  So, in anticipation, I am setting the ground
work now.