Added the ability to subset global properties.
I didn't want the situation where there are a lot of small property
files all over the place and then playing that game where  you stack
up the reading of property files in "just the right order" to get
desired effect.  The goal here is to have a *single* property file for
the application.  However, since the application constists of
individual components I am providing a way for each component to get
the specific set of properties that it needs.  So we get the
partitioning of properties without the undue managment of many
property files.  The only thing to "worry" about is naming collisions
and schemes, but since all the properties are right there to see it
should be obvious what to name your properties :-).