Cleaner order of operations (initializations) and added Resource loading!
There was a bit of inconsistent treatment of the DataNodeService
w.r.t. it being treated as a component, yet doing a bit too much and
stepping on the toes of the ESGDataNodeManager.  So that has been
fixed by having the service essentially be responsible for "booting"
the system, just enough to then play it's roll as any other component.

Initialization of components are now done as a part of registration.
When a component registers itself with the datanode manager - that is
when it's init() function is called.

Added a brand spanking new Resource loading class to core.  Well,
actually this class is rather old school (some code I dug up somewhere
from back in the day) but it does the job.  Now we can read in
configuration files etc from the classpath without having to worry
about the serlet container hitting us on our knuckles. :-)