Attempt to "classify" Events using the class typing system to help
Ideally events themselves would be of a single type and what dictates
the use and routing is what is inside.  This is an idealized model of
the system.  The truth is that we should take advantange of the typing
system to help enforce code / semantic correctness.  This will become
more apparent as we will take advantage of generics in typing event
handlers to the events that they  handle.  I digress.  Pardon...

The java typing system is being leveraged to segregate types of
events.  Though instanceof is a slow operation I feel we should at
least have a code level of making sure events are of the right kind.
This is a foray into starting to establish a "system" type of message
that would supercede other directives (other events) to perform
"interrupted" system tasks (at the framework level).

(I talk too much)

:-)