Notification is in place (mostly)...
Put in notification sub system (component).  Using javamail to perform
email notifications.  At the moment have stubbed out the database
access object (more on that to come). This portion takes care of the
"sweep" in the "mark-n-sweep" process we will use for notification.
The "mark" portion will be performed by future development of code
that puts data into the database that refers to changes in the state
of data.  Also we will collect data with regards to who is using
certain bits of data to know who to notifiy with this component :-).

First cut is good.  Uses an email message template file loaded as a
Resource and token replaced using java's regex.  Runs on a timer and
triggered by events to indicate that there is a state change and to
send notifications.