Initial Database Integration Step
Created the DatabaseResource class.  This class is not a "component",
it is responsible for setting up the database as a resource for the
system.  Namely it loads the driver and establishes a
PoolingDataSource for database connectivity.  The object is a
singleton initialized by the AbstractDataNodeManager .

Added appropriate database properties to the datanode.properties file.

The ESGNotifier component is pretty much all set to go. (It has been
tested with a stubbed DAO and successfully sends email notifications).
Now uses the new NotificationDAO constructor that takes a
datasourece.

The only thing missing at the moment are the queries to get the
notification information. And the code for handling the results
You know... database stuff... Thus far it all builds! :-)

Using commons dbcp, dbutils, pool libraries directly (not counting on
containers or anything of the sort).