DB: Think about the data - innocuous changes...
The changes are not a real issue here... The issue is that I have to
give a bit more thought with regards to the information exchange
between the publisher and the notification component.  At the moment I
am of the mind that the publisher should probably send a list of
dataset ids that have been updated.