Service Filter for THREDDS et. al...
Built filter for logging data access.  The idea here is to log fetches
for data done via the thredds web service.  This will provide the
information that will be referenced by the Notificaton component.
Since datasets are the fundamental unit of the system it would be best
to get a list of the datasets that have changed because of an update
(change in it's constituent files).  Is there a way to get that
dataset information from the incoming url?  If so that is what I
should be recording along with everything else.

Work in progress all stubbed out and compiles (still needs testing).