Further distribution preparation.
Tokens are used for mail properties, that were once hard coded.  Now
there is the ability to turn on and off token filtering (replacement)
by using the property -Dfilter=[true|false] the default is false which
means don't do any filtering / token replacement.  This is the desired
way to build a general distribution. There is also the make_dist
target that will do the task of building the esg-node.#.#.#.tar.gz
distribution file.

:-)