Notification Database grafting closer... still in flux...
Added method registerWithNotificationRunLog to test if a record exists
for this particular data node, identified by a node id (to be gotten
from the bootstrapping of the node - TBD).  If there is no record for
this node's id then one is created.  From then on when a notification
is set it updates the time field (seconds since epoch) corrensponding
to this node's id.  This provides multi node support using the same
database.  Though this is not the usual use case, as we strive to
stove-pipe each data node installation.  It is only a small bit of
logic to allow for a more flexible configuration, so why not. :-)
(the case where a database install is shared among different data
nodes).