NotificationDAO setting default node ID if one is not set.
Setting the nodeID to the ip address of the local host if one is not
explicitly set.  This logic is here, for now... but will probably move
deeper into the core as ... yeah, data nodes should have ids :-)...
In this context the IP address is a good id.  to support single
database, multi manager configurations. (though not the recommended
scenario).