Notification service sends email notifications!
The query for notification and the sending email as first (and
currently only) notification mode is complete.

I have documented in the class comments what I would do to make this
process more memory efficient/friendly if it is seen that the memory
footprint under notificiation is too large.

NOTE:
There is also a side issue where I would like to put in some code to
reset the datasource pool if the database goes down.  This would mean
that you don't have to restart the application in order to
re-establish a connection to the database if the database goes down.
(something to think about later)