Much improved "post" script
The post script is a simple utility to push files from the tree to the
distribution server.  It now is more robust at reading the files.
Also it can read "ini" style files, where the destination directory
can be given within []s and can be individually overridden by
specifying an additional destination token after the file entry.
See the post script's header comments for details.  If run with
"debug=1" variable set, the output will be more robust.

It still creates the md5 files for each file posted. :-)

It is a small little thing, but I really like it!  Go BASH!