Double check downloads for idempotent update behavior
Added the check_for_update function and have it being called by the
checked_get function.  The result is that before a file is downloaded
from the server it's md5 is generated and checked against the
associated md5 file posted on the server.  If this is a match then
nothing further, and the checked_get function returns with a 1 - not
an error but just an indication that checked_get did not get anything
because the file had not changed.  If the return value is > 1 then
there was something to download but there was an error trying to get
it OR the file was downloaded but then what was downloaded does not
match the posted md5 for that file... i.e. the file cannot be
"verified".  Files that cannot be verified should not be used!!!

This functionality is the other side of the "post" script that I
described in the previous commit.

Just making sure that this part of the dev cycle is tight!

-Gavin :-)  Friendly Neighborhood BASH monster.