Added esg-update script...
Added the script, update script to the project.  This script updates
the user's machine with the other scripts used for the install ex;
esg-node.  It also uses the new checked_get check_for_update
functions. The only thing the user has to do is manually check the md5
on this file (boostrapping issue :-).  All subsequent file pulls are
internally md5'ed.