Including esgcet in --upgrade
Added the upgrade flag to setup_esgcet and added the call to
setup_esgcet into the --upgrade flag switch.  The upgrade of the
esgcet code runs only the upgrade of the install... NOT THE DB.  When
the upgrade flag is set (1) then the function returns 0 right after
the call to upgrade.

The upgrade for esg-node follows.