Metrics and Monitoring stubbed out...
Put in the basic framework code for Metrics and Monitoring, including
property values in the property file and sewing them into the
ESGDataNodeManager.

(Next?)
Have to create and install the DAO for the monitoring.
Have to create and install the system calls for metrics.
Have to think about how to establish caching (though don't want to
optimize too soon)