Monitoring Component Scafolding... (no queries yet)
Monitoring component that provides statistics on the system.  In
particular statistics based own popularity as indicated by downloads.
In doing this created a small hierarchy for Data Access Objects
(DAO's) so that doing this is more boiler plate.  At the moment things
are setup for a 1:1 between DAOs and select queries.  Encapsulated
result objects are member classes of the DAO.  The DAO's are
aggregated and called by the ESGMonitor code.