Metrics and Monitoring RE-Stubbed out...
Metrics: For updating statistics about the data node
Monitoring: For updating state of host and processes.

I had there implementations reversed so now it is sorted out.  The
queries still need to be put in place to actually collect the
necessary data. Once the queries are in place, the output result
objects need to be populated and send to the caller.

NOTE: There will have to be a caching system put in place (but now
right now).