Modifications to esg-node script to support GridFTP configurations
Until now the GridFTP (token based) configuration that has been in
place provides additional token based facilities for ESG end-users. A
different configuration is require for GridFTP when used by the Bulk
Data Mover (BDM).  So now the script is modified to pass args to the
esg-globus script functions to direct what kind of GridFTP
configuration should be done.

There is an additional switch in the script:
"--gridftp-config"
which takes one of two args:
"with-bdm-config" | "bdm-config-only"

Note: should probably be nice with verbiage by taking out "config"
from the arg values.  (something to do later when I am bored).

BTW: if you don't use the switch at all the default is the "end-user"
configuration.  You can specify the values by setting the environment
variable "gridftp_config_args" to either of the above values.

:-)