DiskInfo Working...
Got the regular expression and such working for pulling the dataset
roots out of the Publisher's ~/.esgcet/esg.ini file.  Then it collects
the information on each disk listed.  The information made available
does not include the filesystem path just the dataset root given
names.

P.S.
Pretty sexy using the mmapped file... In this case not so much, but
for the other system monitoring calls that will be made available is
subsequent commits, I'll be reading right against the /proc pseudo
filesystem and mmapping that info... (at least that is the plan, let
me not get too excited and speak too soon :-)