Almost finished with System Stats (monitoring) code
There is a kernel bug in mmapping the /proc/meminfo file such that the
first two lines get garbled.  This is apparently a known bug that at
the moment I have not figured / learned a way to surmount. (dag).

/proc/meminfo looks like this...

40 kB
DirectMap2M:     2086912 kB
e:          449360 kB
Buffers:         1144228 kB
Cached:           223920 kB
(yadda yadda yadda)

Also created properties for tuning the buffer sizes so you can read
the /proc files in one big 'gulp' for efficiency.  Not closing
buffers, also for efficiency.

This is well tuned (sort of)... the RegEx do add to the effort, though
they are fast, and the conversion from bytes to char may be biting me
in teh butt, but thus far have timed about 5K reads a second. which is
sure to degrade with all the other stuff piled on, but this should be
PLENTY good for a every 30 second update and cache. right? :-)