Fleshed out ESGMonitor and ESGMetrics DAOs.
Lazy fetching of data and caching results. Queries are now sewed in
and should be working.  Still have to do some cache implementing which
will in clude winding down caches so that they get cleared after being
inactive for a certain amount of time.  Also have to put in timeseries
monitor results collection.