Factored out Hessian specific dependency
Created HessianGateway object to encapsulate the Hessian factory
call.  Wrapped exceptions with ESGException appropriately.  This is a
first level factoring based on having only a single RPC package in
hand so... it is nascent.

(stubbed a silly little test main program for metrics have to do work
there, but decided to commit it anyway)