Added SEDA style event queuing to framework.
As it was before there is a composable framework of components that
register with the component manager and thus are able to get and fire
"system" events.  For example; these types of events allows components
to know when other components have joined or left the overall set of
components that make up the system, as well as when the system had
loaded all components.  Furthermore since all components register with
a central manager of which they have a handle to, they can use this
manager to get handles (by name) to specific desired components they
are "interested" in.

Along with that "control" event layer, I have just added a SEDA style
queue to each component.  This allows the system to handle multiple
threads by way of multithreded queues and provides the "event-driven
concurrencey" we desire.  The use of the queues and event based
interface gives the system loose coupling between components ('stages'
in SEDA parlance) - "which improves modularity and robustness".

One of the issues with such a message-oriented model surrounding an
event queue is the "no-free-lunch" aspect of having many, short-lived
threads enqueing and dequeuing events and the context switching they
incur. This issue is ameliorated by the current trend in computing
where we have machines with many many cores and hence less and less
need for having to do context switching :-).  Indeed managing long
threads throughout a program is the dual of handling short thredds
between stages... but the latter is easier to program, extend and
monitor.

See the SEDA paper
http://www.eecs.harvard.edu/~mdw/papers/seda-sosp01.pdf
http://www.eecs.harvard.edu/~mdw/proj/seda/

In conclusion this is the very beginning of having a mix of OSGI and
SEDA style methodologies to create a robust, composable, framework.
(To be honest I was coding this all up before I even heard of OSGI but
since I stumbled upon it - it seems close to what I have).  They key
is I want to make this as simple as possible.  Any time I see
projectds like the OSGI and I look at the java doc and see tons and
tons of classes, I feel that somehow they are just overdoing it just a
little bit.  Same some code for the programmers! :-)  Anyway, my
commit messages are too damn long!

TODO:

The pool and batch controllers are stubbed out classes and I still
have to decide on the best way to connect the controller to the
manager in the scenario for having a Uber controller manager.