Preserves node.properties file in a more clean way
The best solution would be to have a the node.properties file be
called node.properties.template from the very beginning and then
during installation if there is not already a node.properties there
then copy the node.properties.template over to node.properties and
peform the token substitution.  This would be the best way...
However, I think there is a dependency with the build.xml file that
may be an issue.  I have to take a closer look at it... but for now
this way works and is more clean that then last checkin by keeping
everything at least in the same WEB-INF/classes tree.

NOTE: Revisit this when you have some tinker time.