Lots of changes: Put in message Queues.
So basically the SEDA style Queues are in place and the event routing
is in place. This still needs a lot of testing but at the moment
things don't crap all over themselves so I figured it would be a good
time to commit some stuff.

TODO:
Fix the Basic Gateway's isAvailable check...
Add the ESGMonitor time series collecting of status...
Add the LRU cache for the time servies collecing....
Make messages for remote calls to Metrics and Monitoring...
Put in IDs for remote events to better support book keeping for async
calls.

lots of other stuff... (tired... going home now)