Product upgrade symlink handling / bettern account group handling
Updated the version of ant being installed.  Now have coded up the
installation such that when new versions of ant or tomcat are
specified the old versions get unlinked and/or moved out of the way to
make room for the new version and re-symlinking.

Since I had put in the ability to specify different group values for
users that this script creates I also had to now put in the
modification such that we make sure the group exists or is created
before the user is created.

Since the /etc/esg.env file is read and the values there usurp local
environment variables, I made it so that you can call --clear along
with the other options to clear out that file before
installation/upgrading begins.  As a side effect this means that the
script will run with all the script defaults!

TODO: need to put in something to clean out duplicate keys in the
/etc/esg.env file (keeping the lasts).