esg-update script will pull esg-node script from diff location
Basically now we pull the node script from the esg-node directory
rather than from externals/bootstrap.  This is part of a small
reorganization that reflects the fact that the esg-node is not a
project "external" to PCMDI, whereas the other codebase is (eventhough
I wrote the install scripts for them as well).