Subsystem installs script mechanism - Security (sub)script called.
Set up main esg-node script to be able to run "sub"system scripts
"Here we are enforcing a bit of a convention... The name of subsystem
files must be in the form of esg-xxx-xxx where the script contains its
"main" function named setup_xxx_xxx(). The string passed to this
function is "xxx-xxx"" Also an optinal directory for where the file is
on the distribution site.