esg-node: Lots of structural changes...
So I re-ordered the physical location of the functions to match the
order that they are called.  Also took out the setup_filter code and
put it into the filter sub-system scripts over in bootstrap.

Thredds installation is done "manually".  tomcat is stopped and then
the installation takes place, the rest of the configuration takes
place and then tomcat is restarted.

Made input loop around the selection of gateway endpoint.  Also now
users can manually input their own endpoint with the [C|c] option.