Clean up: No longer hard setting the gateway endpoint for the basic gateway.
Uses the default.gateway.service.root property to set the default
gateway. During installation this value is the value of
ESG_GATEWAY_SVC_ROOT. this value is written into the node.properties
file. IMPORTANT TO NOTE: Once it is set in this file it is used
regardless of what the environment ESG_GATEWAY_SVC_ROOT!
(perhaps I should read that environment variable instead of only from
the node.properties file... for another commit, perhaps)