DataNodeManager will now respect either the node.properties file's default.gateway.service.root property or will use the environment's ESG_GATEWAY_SVC_ROOT value.
The latter is used to create the former, however, there may be times
when we want to make the changing of endpoints more fluid (easy).  If
the property exists (not null) then it will be used... otherwise the
environment variable if set will be used... otherwise there will be a
null pointer exception that will break the system!

(node: the system will have to re-register with the new endpoint)

during installation.