esg-node: Now java is installed as well
Want to make sure that folks have the "right" java, since our
application does not work with the open java et. al.  The only thing
of note here is that WE at PCMDI are holding the java distribution
which is a jdk 1.6.0_20 exploded then re-tarred.  It is for i586 linux
machines.  In general we want people to have their own java installed
- this is a nicety that we are offering especially since we have run
into issues with java versions and jsse interactions with ssl
handshaking.