esg-node: Script now supports versions or at least is more version aware
Augmented check_version function so that it can read version values
from java, ant, git and postgress.  I pulled the out each of those
programs into variables put at the top of the script.  The "setup"
method for each of these products checks the version and if the
version it detects is too old (i.e. check_version returns != 0) then
the rest of the setup function will execute.  It still asks, but now
it won't ask if you have the appropriate version installed.

check_version() only works with programs that respond to the
(-)-version flag.  Notice: must test "--" before "-" so that
postgresql can be included (else postgress wants you to login, yadda
yadda yadda).  So now we have somewhat of a version / change control
mechanism in this script.

No joy with tomcat, thredds (war), cdms (python mod) and esgcet
(publisher).  However esgcet (the publisher) has python to take care
of its versioning.

This is a first step... :-)

God bless UNIX, BASH, and REGEXP!!!

P.S. - I had no idea this script would be so feature rich or LONG!
But damn it's useful and pretty fresh, if I do say so myself.