esg-node: fixing problem with myproxy...
removing the certificate directory entirely if it exists.
I would have set it to X509_CERT_DIR, but then that would be a huge
security problem telling a script to nuke a directory specified by a
variable.  Right... So fixed that var to the value that X509_CERT_DIR
should be set to ~/.globus/certificates-esg