esg-node: jssecacerts moved to java's jre/lib/security created start() stop() functions to facilitate restart flag
For some reason, which is pretty darn sloppy in my opinion... the
jssecacerts file that is used as the truststoreFile in
/usr/local/tomcat/conf/server.xml for the ssl connector needs to be in
$JAVA_HOME/jre/lib/security in order for things to work properly.
Meaning for you to be able to download an NC file from thredds this
must be set appropriately in that place.  It remains to be seen if I
have to take the truststoreFile entry out of the server.xml file
entirely.  Though the latter is the way it should be done!