esg-node: Building Curl now... and some clean up.
Curl is needed by GIT in order to work over the HTTP protocol.  At the
moment this the protocol that must be used in order to fetch the
latest CDAT from the GIT repo.  Hopefully they powers that be here at
LLNL will open up the git port so we can use the must faster git
protocol!

:-)