Stopping Tomcat Less Lethal...
I made it so that stopping tomcat is not so lethal.  It will only do a
kill on the tomcat whos PID is in the
(tomcat_pid_file=/var/run/tomcat-jsvc.pid) file.  I commented out the
section where it just does a killall on all jsvc processes.

Thanks for bringing it to my attention Eric ;-)

This should do it...
-gmb