Added TODO item -> need to setup DB version control like liquibase...
Need to have a clean migration path for upgrades to the datanode
database.  Right now there is none and changes to the database are not
possible without a lot of collateral damage, which means loosing
data.  So putting in something like liquibase will solve that
problem.  Though I don't anticipate the database changing much... it
is good to have!!!!  I hope to get to this soon.

-gmb