Remote Accsss Log fetching with bounds
On the client side, I am properly setting the limit of the time
paremeter to 0 -> Long.MAX_VALUE.  On the server side I am setting the
limit on the query to be 999 records.  This way if the caller gets 999
records they can presume that there may be more records to fetch and
thus can setup sebsequent calls to fetch that data as well.  This puts
"sane" limits on how much memory we are using to fetch and store the
database records and then put them on the wire.  The value of 999 is
arbitrary but I think a reasonable starting point.

(also have the install script esg-node show the service list in the
--update case)