Retooled Code to reference Peers not Gateways.
Essentially going through and getting rid of all references to
"Gateways" and now refer to "Peers".  This meant renaming some classes
and quite a bit of careful replacement.

Also did work with cleaning up the the registration / notification
circuit.  Now when ingress registration calls are gotten we create a
wrapper delegate "stub" object (BasicPeer) that we then register with
the data node manager.  Registering the stub with the DNM creates the
endpoint (in the init call), and then sends a join event out.  The
connection manager gets that join event with the stub inside and puts
it in its local data structure.  Then it calls that stub's
notifyToPeer method.  This method makes the RPC call to the peer that
was registering and calls its notify method and passes it a remote
event with the messgae type NOTIFY.

TODO: Work out initiating the registration.  When should that happen
outside of the bootstrap case.  That code should go in the connection
manager.