Just a comment to make it clear what the 2nd, optional, arg means.
The register bash function takes potentially two args the first is the
host endpoint from which you want to extract the public cert the other
is the password to the local truststore that you will be inserting
this key into.  The 2nd is optional, where there is a default already
set.