Added respect for CATALINA_OPTS calling it ${tomcat_opts} also setting -Djava.awt.headless=true
Estani sent me an email and wanted the script to respect this option.

"Regarding Hudson it is pretty simple to install, but tricky to start...
if you happen (like me) to set it within a tomcat installation, then it
becomes "really" tricky. Sadly the use the AWT library which intends to
access the display. In order to prevent it from doing that, you should
start tomcat with this JVM option:

-Djava.awt.headless=true

I had to modify the esg-node script so that I start tomcat correctly
(yes, I'm running hudson in our ESG backup system :-). A "nice to have"
feature of the esg-node tomcat start script will be to respect the
environment variable $CATALINA_OPTS (If you change that make sure you do
it at the command, the first time I just changed the echoed string :-P)"

So there yah go!