Created esg_root_dir et. al. and Can build jsvc under "new" 6.0.29 commons-daemon
It is clear to me that with the advent of the gridftp usage logging
tool and perhaps more esg specific tools to come that we should give
serious consideration for a basic filesystem structure to accommodate
these kinds of things.

Also, as I was upgrading from tomcat 6.0.26 I noticed that the jsvc
mechanism had changed to using apache commons-daemon library.  So now
the script will accommodate this "new"-er way of building jsvc (for
unix).  In both cases the jsvc binary gets put in the
tomcat_install_dir/bin directory, this way everything still works as
before with minimal impact to the rest of the script.