Setting ownership of specific directories property - so publisher can work
Getting the user ID from the $HOME environment variable.

Using it to chown the $HOME/.esgcet and the
$CATALINA_HOME/content/thredds/esgcet to the "correct" person.  This
means that the publisher is also the same as the installer (sort of).
We may want to relax this and just chmod 777 these directories.  For
now we will change ownership and see how far that gets us.  If there
is a demand for allowing publishing from more than root or that user,
then we can easily adjust this.

-gavin