Added pattern matching for url file extensions.
Now we only log request urls that match a certain pattern extension.
In this case we are looking for .nc files.  In the future we could
easily make this a parameter that gets passed in such that the xml
descriptor would be the place to enter these values. At the moment
leaving it hard coded is good enough.