Log web access to files with certain extensions specified in parameter
So now for the AccessLoggingFilter I make sure that it only logs http
requests that are for specific file extensions.  Now these extensions
can be specified in the web.xml filter parameters.

<param-name>extensions</param-name>
<param-value>.nc,.foo,.bar</param-value>

I have hard coded the ".nc" extension and have done a bit of defensive
coding for nulls, etc.