Made AccessLoggingFilter code match database schema
There is a field called "batch_update_time" that was introduced that I
cannot figure out for the life of me why it is there and what its use
is.  Apparently this came out of some discussion that I was perhaps
"out to lunch" on.  But somehow this field made it's way into the
schema (a while ago) and now I have made the code match the call.

I suspect I will launch a campaign to remove this field but for
now... We'll let it ride.

I really have to put in liquibase or something so that database schema
changes aren't so painful.  Oh well.

(I have to admit I did a small format change s/tab/space(s)/g that
confounds this checking from being prestine, my bad)