Fixed sql bugs - no mismatch stored proc params - no mega intersect on egress
Broke up the "ingress" query so that I make the sql call to get the
primary key, then I use that value to add the entry into the database.
Breaking it up allows me to hold the value of the primary key in the
calling environment and return it if the query succeeds.  The calling
environment now uses that primary key value that it holds to update
the inserted record on ingress with status information on egress (the
success of fetching the file and the "duration");

I still don't believe that Nate is right with respect to finding the
duration of a download.  It looks like to me, all duration does is
tell you how long it took to traverse down the filter chain and bubble
back up again from the call you made.  I think I need something that
is where the bits are being pumped so I can count how many have gone
through and if that succeeded.  This is a TODO for another time.

Also changed logging levels to debug for debug-like output

-gmb