The default behavior should be to run what you have [c] if a content skew is detected.
if a change in hash value is detected.  The right thing to do would be
to pull down the cleanest version from the source.  However, to be
even more paranoid and essentially not trust the source, we should
just do nothing until everything can be fully verified.  But to be
practical, we don't want people's systems not coming up because of an
update detection... and since it is hard to do an in-place update of
the same executing code (someone convince me otherwise), then the
thing to do is to continue with what you have.

We don't want to make it cumpolsory that you upgrade your script.
Being sesitive to my man Ross' position, if it ain't broke, I don't
trust to upgrade it! LOL.  Indeed, so now by default your wishes are
taken into account.

-gmb