Made things simpler... Don't have to know file extensions...
Okay so now I put the file extensions used for compressed files in a
variable.  I return to using awk to use the gsub method in order to
truncate files appropriately.  This way I don't have to explicitly
state the file extention, it just has to be included in the global
variable $compress_extensions.  This gives the script a little less
maintenance burden, though you still have to know the correct command
to uncompress the file.  Makes things a bit more uniform.

-gmb