Keystore Password put back in global scope: This must change!
The esg-security sub-script that gets called looks for this value to
replace in its application.property file.  I don't like having this
value out at script scope level because it can be picked off.  There
is a non-zero security issue here that I'd rather not expose.

I will have to modify the ORP's esg-security sub-script to parse the
server.xml file to pull out the credentials from there.  This also
makes this more portable in the long run.  Actually what would be even
better is to have the code itself pull this value, then it can be more
dynamic to changes to the keystore so it always has the right value.
-gmb